// src/enums/actions.enum.ts
export enum ActionType {
  SET_COLOR = 'SET_COLOR',
  SET_CANVAS_SIZE = 'SET_CANVAS_SIZE',
  UPDATE_DRAW_COLOR = 'UPDATE_DRAW_COLOR',
  UPDATE_CELL_SIZE = 'UPDATE_CELL_SIZE',
  UPDATE_BRUSH_SIZE = 'UPDATE_BRUSH_SIZE',
  MOVE_MOUSE = 'MOVE_MOUSE',
  MOVE_PRESSED_MOUSE = 'MOVE_PRESSED_MOUSE',
  PUSH_HISTORY = 'PUSH_HISTORY',
  MOVE_BACK_IN_HISTORY = 'MOVE_BACK_IN_HISTORY',
  MOVE_FORWARD_IN_HISTORY = 'MOVE_FORWARD_IN_HISTORY',
};